pyproject.toml
[project]
name = "slop-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.107.3",
    "python-dotenv>=1.1.1",
    "python-whois>=0.8.0",
    "aiohttp>=3.9.0",
    "tabulate>=0.9.0",
    "tqdm>=4.67.1",
]
